% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_census_blocks.R
\name{sample_ward}
\alias{sample_ward}
\title{sample locations within a ward}
\usage{
sample_ward(x, n, coord_names, covar_names, seed = NULL)
}
\arguments{
\item{x}{A data.frame of the population you wish to sample from, where each
row is a possible sampling location.}

\item{n}{The number of points to sample}

\item{coord_names}{Either a numeric vector that describes the columns of the
coordinates you wish to include as auxilary variables
from the data.frame x or a character vector of the
column names.}

\item{covar_names}{Either a numeric vector that describes the columns of the
covariates you wish to include as auxilary variables
from the data.frame x or a character vector of the
column names.}

\item{seed}{An optional value to set the seed for sampling. If NULL, this
will be generated for you.}
}
\value{
A named list. The first object 'row_indices' indicates the rows
           from data.frame x that represent a balanced sample. The second
           object 'seed' is the seed used to generate the balanced sample.
}
\description{
\code{sample_ward} generates a spatially balanced sample.
}
\examples{

\dontrun{
  data <- read.csv(
    "./data/my_ward_data.csv",
    stringsAsFactors = FALSE
  )

  my_samples <- sample_ward(
    x = data,
    n = 12,
    coord_df = c("lat", "lon"),
    covar_names = c("medincome", "ratcomplaints", "popdens")
    )
}

}
